Remove combineReducers shortcut in favor of an explicit call
