+++ examples/todomvc/containers/App.js
@@ -1,10 +1,11 @@
 import React from 'react';
 import TodoApp from './TodoApp';
-import { createStore } from 'redux';
+import { createStore, combineReducers } from 'redux';
 import { Provider } from 'react-redux';
 import * as reducers from '../reducers';
 
-const store = createStore(reducers);
+const reducer = combineReducers(reducers);
+const store = createStore(reducer);
 
 export default class App {
   render() {
